test-automation/LmsApi/LmsFrontend/requirements.txt
configparser==5.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.28
pandas==1.4.3
pytest==7.1.2
pytest-forked==1.4.0
pytest-html==2.0.1
pytest-html-reporter==0.2.9
pytest-metadata==2.0.2
pytest-ordering==0.6
pytest-reportportal==5.1.3
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
requests==2.25.1
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
selenium==4.7.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
urllib3==1.26.12
virtualenv==20.0.17
webdriver-manager==3.8.5
